+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Infer default namespace from kubeconfig when not configured via the provider (https://github.com/pulumi/pulumi-kubernetes/pull/1896)
+- Fix an error handling bug in await logic (https://github.com/pulumi/pulumi-kubernetes/pull/1899)
 
 ## 3.15.1 (February 2, 2022)
 - [Helm/Release] Add import docs (https://github.com/pulumi/pulumi-kubernetes/pull/1893)